Separate `pow` & `powf` in `PowOps`
* `pow(&self, n: Self) -> Self`
* `powf(&self, f: f64) -> Self`

+++ examples/dual.rs
@@ -6,7 +6,7 @@ fn main() {
     let m = dual(2, 1);
     (1f64 - m / r).print();
 
-    m.powf(r).print(); // (x^3, 3x^2) = (8, 12)
-    r.powf(m).print(); // (3^x, 3^x * ln(3)) = (9, 9.8875...)
-    m.powf(m).print(); // (x^x, x^x(ln(x) + 1)) = (4, 6.7725...)
+    m.pow(r).print(); // (x^3, 3x^2) = (8, 12)
+    r.pow(m).print(); // (3^x, 3^x * ln(3)) = (9, 9.8875...)
+    m.pow(m).print(); // (x^x, x^x(ln(x) + 1)) = (4, 6.7725...)
 }
